requirements.txt
# --- Google Cloud & Generative AI ---
google-cloud-bigquery==3.12.0
google-genai==0.3.0
google-cloud-aiplatform==1.35.0
db-dtypes==1.1.1 
google-cloud-discoveryengine

# --- Data Processing & Analysis ---
pandas==2.1.1
numpy<2.0.0
openpyxl==3.1.2                  # Required for Excel operations
lxml==4.9.3
Pillow                      

# --- Networking & Web Scraping ---
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1 

# --- Environment & Utilities ---
python-dotenv==1.0.0